+++ packages/client/src/feature/forum/ui/AddMessage/AddMessage.module.scss
@@ -1,7 +1,7 @@
 .root {
   display: flex;
   flex-direction: column;
-  align-items: end;
+  align-items: flex-end;
   padding: var(--s16);
 }
