+++ packages/client/src/feature/forum/ui/AddMessage/AddMessage.tsx
@@ -1,60 +1,30 @@
-/* eslint-disable react/require-default-props */
-import { Dispatch, FC, SetStateAction, useState } from 'react'
-
-import { useDispatch, useSelector } from 'react-redux'
+import { FC, Dispatch, SetStateAction } from 'react'
 
 import BaseBox from '@/shared/components/ui/BaseBox'
 import BaseTypography from '@/shared/components/ui/BaseTypography'
 import BaseTextField from '@/shared/components/ui/BaseTextField'
 import BaseButton from '@/shared/components/ui/BaseButton'
 
-import { addMessage } from '@/entities/message/model'
-
-import { getMessages } from '@/entities/message/model/selector'
-
-import formatDate from '@/shared/helpers/formatISODate'
-
 import styles from './AddMessage.module.scss'
+import useMessage from '../../hooks/useMessage'
 
 interface Props {
+  // eslint-disable-next-line react/require-default-props
   replyToId?: number | null
   setReplyToId: Dispatch<SetStateAction<number | null>>
 }
 
-const AddMessage: FC<Props> = props => {
-  const { replyToId = null, setReplyToId } = props
+const AddMessage: FC<Props> = ({ replyToId = null, setReplyToId }) => {
+  const { newMessage, messagesData, handleAddMessage, handleChange } = useMessage(replyToId, setReplyToId)
 
-  const [newMessage, setNewMessage] = useState<string>('')
-
-  const messagesData = useSelector(getMessages)
-
-  const dispatch = useDispatch()
-
-  const handleAddMessage = () => {
-    dispatch(
-      addMessage({
-        id: 6,
-        autor: 'asd',
-        date: formatDate(new Date().toISOString()),
-        message: newMessage,
-        replyToId: replyToId || undefined
-      })
-    )
-
-    setNewMessage('')
-    setReplyToId(null)
-  }
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setNewMessage(event.target.value)
-  }
+  const replyiedMessage = messagesData.find(m => m.id === replyToId)?.message
 
   return (
     <BaseBox className={styles.root}>
       {replyToId && (
         <BaseBox className={styles.reply}>
           <BaseTypography variant="body2" color="textSecondary" gutterBottom>
-            Ответ на: {messagesData.find(m => m.id === replyToId)?.message}
+            Ответ на: {replyiedMessage}
           </BaseTypography>
         </BaseBox>
       )}
