+++ packages/client/src/shared/config/routing.ts
@@ -7,7 +7,6 @@ export const paths = {
   leaderboard: '/leaderboard',
   forum: '/forum',
   forumTopic: '/forumTopic',
-  newForumTopic: 'newForumTopic',
   page404: '*',
   page500: '/500'
 }
