+++ packages/client/src/widgets/forum/ui/MessageList/MessageList.module.scss
@@ -3,7 +3,7 @@
 }
 
 .paper {
-  margin-bottom: var(--m20);
+  margin-bottom: var(--s20);
   padding: 0;
 }
